nvim: initialise vimwiki
I use vim-wiki to make notes now.
In future, I would like to write a filesystem for it based on
StandardFile. This would make it easier to edit notes on-the-go and make
it easier to sync all around.
VimWiki would be used as a means of writing and exporting to HTML for
on-the-go viewing experience.

Signed-off-by: Gunwant Jain <mail@wantguns.dev>